+++ src/tcNode/Properties/Resources.Designer.cs
@@ -619,9 +619,9 @@ namespace TradeControl.Node.Properties {
         ///   Looks up a localized string similar to /**************************************************************************************
         ///Trade Control
         ///Upgrade script
-        ///Release: 3.30.3
+        ///Release: 3.29.4
         ///
-        ///Date: 24 September 2020
+        ///Date: 14 September 2020
         ///Author: IAM
         ///
         ///Trade Control by Trade Control Ltd is licensed under GNU General Public License v3.0. 
@@ -699,6 +699,36 @@ namespace TradeControl.Node.Properties {
         internal static string tc_upgrade_3_30_2 {
             get {
                 return ResourceManager.GetString("tc_upgrade_3_30_2", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to /**************************************************************************************
+        ///Trade Control
+        ///Upgrade script
+        ///Release: 3.30.3
+        ///
+        ///Date: 1 October 2020
+        ///Author: IAM
+        ///
+        ///Trade Control by Trade Control Ltd is licensed under GNU General Public License v3.0. 
+        ///
+        ///You may obtain a copy of the License at
+        ///
+        ///	https://www.gnu.org/licenses/gpl-3.0.en.html
+        ///
+        ///Change log:
+        ///
+        ///	https://github.com/tradecontrol/tc-nodecore
+        ///
+        ///Instructions:
+        ///This script should be applied by the Node Configuration app.
+        ///
+        ///************ [rest of string was truncated]&quot;;.
+        /// </summary>
+        internal static string tc_upgrade_3_30_3 {
+            get {
+                return ResourceManager.GetString("tc_upgrade_3_30_3", resourceCulture);
             }
         }
     }