+++ src/tcNode/Properties/Settings.Designer.cs
@@ -96,6 +96,7 @@ namespace TradeControl.Node.Properties {
   <string>tc_upgrade_3_29_4</string>
   <string>tc_upgrade_3_30_1</string>
   <string>tc_upgrade_3_30_2</string>
+  <string>tc_upgrade_3_30_3</string>
 </ArrayOfString>")]
         public global::System.Collections.Specialized.StringCollection SqlScripts {
             get {