+++ src/tcNodeDb/Cash/Views/vwTaxCorpTotalsByPeriod.sql
@@ -8,6 +8,7 @@ AS
 								 App.vwCorpTaxCashCodes CashCodes  ON Invoice.tbTask.CashCode = CashCodes.CashCode INNER JOIN
 								 Invoice.tbInvoice ON Invoice.tbTask.InvoiceNumber = Invoice.tbInvoice.InvoiceNumber INNER JOIN
 								 Invoice.tbType ON Invoice.tbInvoice.InvoiceTypeCode = Invoice.tbType.InvoiceTypeCode
+		WHERE CashTypeCode = 0
 	), invoiced_items AS
 	(
 		SELECT (SELECT TOP (1) StartOn FROM App.tbYearPeriod WHERE (StartOn <= Invoice.tbInvoice.InvoicedOn) ORDER BY StartOn DESC) AS StartOn,  
@@ -16,11 +17,32 @@ AS
 							  App.vwCorpTaxCashCodes CashCodes ON Invoice.tbItem.CashCode = CashCodes.CashCode INNER JOIN
 							  Invoice.tbInvoice ON Invoice.tbItem.InvoiceNumber = Invoice.tbInvoice.InvoiceNumber INNER JOIN
 							  Invoice.tbType ON Invoice.tbInvoice.InvoiceTypeCode = Invoice.tbType.InvoiceTypeCode
+		WHERE CashTypeCode = 0
+	), assets AS
+	(
+		SELECT cash_codes.CashCode, financial_periods.StartOn, 
+			CASE cash_codes.CashModeCode WHEN 0 THEN financial_periods.InvoiceValue * -1 ELSE financial_periods.InvoiceValue END AssetValue
+		FROM App.vwCorpTaxCashCodes cash_codes
+			JOIN Cash.tbPeriod financial_periods
+				ON cash_codes.CashCode = financial_periods.CashCode
+		WHERE cash_codes.CashTypeCode = 2
 	), netprofits AS	
 	(
-		SELECT StartOn, SUM(InvoiceValue) AS NetProfit FROM invoiced_tasks GROUP BY StartOn
+		SELECT StartOn, SUM(InvoiceValue) NetProfit 
+		FROM invoiced_tasks 
+		GROUP BY StartOn
+		
 		UNION
-		SELECT StartOn, SUM(InvoiceValue) AS NetProfit FROM invoiced_items GROUP BY StartOn
+		
+		SELECT StartOn, SUM(InvoiceValue) NetProfit 
+		FROM invoiced_items 
+		GROUP BY StartOn
+
+		UNION
+
+		SELECT StartOn, SUM(AssetValue) NetProfit
+		FROM assets
+		GROUP BY StartOn
 	)
 	, netprofit_consolidated AS
 	(